new file to test rules new utils.py file to hold stuff used in other files  trying to avoid interdependency
**mem limits are not firing alarms**
